clangd/asm: Generate AST for Assembly files
Prototype of Assembly support in Clangd. The AST is smuggled around on
the existing C-like AST, and can be queried in locations that want to
add ASM support (see later patches). The C AST itself is mostly 'empty',
just providing the bare minimum to not crash immediately. Existing code
should be able to treat it like a very uninteresting file, while
dedicated Assembly aware code paths can query for and work with the
Assembly AST field.

Signed-off-by: Benjamin Gray <bgray@linux.ibm.com>