clangd/asm: Support instruction documentation on hover
One of the main motivations behind this project was because I had no
idea what the 'li' instruction did when I started with PowerPC assembly,
and it was difficult to track down in the documentation. It would have
helped a lot to have editor support telling me something about the
instruction, and where I could find the official documentation.

This patch adds such support to Clangd. Using a 'database' of
instructions given by the `CLANGD_DOCS_PPC` environment variable, it
will load in descriptions for instructions and display the appropriate
content for the hovered instruction.

Ideally the 'see more on page' text can be a link that, when
clicked, would open the documentation to that page automatically. This
is easiest for HTML docs, but is doable with PDF docs too if the client
(or clangd?) can intercept the URL and run the appropriate CLI command
for some configured PDF viewer (Okular, Evince, Chrome, FireFox, etc.).

The database can be generated from the existing documentation, but
I'm not quite sure on the copyright regarding distributing that if it
gets made. So for now I've attached a mockup.

Signed-off-by: Benjamin Gray <bgray@linux.ibm.com>