clangd/asm: Add a mock instruction database for PowerPC
This is an example of an instruction database that can be loaded by
Clangd's ASM support. The contents has been stripped to avoid
reproducing any text of the ISA for now, but in theory it could present
the same content as each instruction's section in the ISA.

Clangd will load it if you set the "CLANGD_PPC_DOCS" environment
variable to the absolute path of this file. Test it on a branch
instruction, like `blr`, by hovering over the instruction name.

Signed-off-by: Benjamin Gray <bgray@linux.ibm.com>